gps計測値をテキストファイルに保存
