+++ GPS/GPS関数管理/main.py
@@ -1,9 +1,20 @@
-from gps import GPS
-#from gps0605 import setupGPS
-#from gps0605 import gpsread
+import time
+import gps 
 
-#setupGPS()
-#gpsread()
-#setupGPS()
-#GPS.setupGPS()
-GPS.gpsread()
+gps = gps.GPS()
+#file kakikomi
+#G = gps.GPS() #self
+gps.setupGPS()
+start_time = time.time()
+
+while True:
+    gps.gpsread()
+    timer = 1000*(time.time() - start_time)
+    timer = int(timer)
+    datalog = str(timer) + ","\
+              + str(gps.Time) + ","\
+              + str(gps.Lat) + ","\
+              + str(gps.Lon)
+    
+    with open("test.txt",mode = 'a') as test:
+        test.write(datalog + '\n')
